Remove the old CPU threading remains, start redesigning interfaces.

+++ ext/native/thin3d/GLQueueRunner.cpp
@@ -889,8 +889,8 @@ void GLQueueRunner::PerformBindFramebufferAsRenderTarget(const GLRStep &pass) {
 
 void GLQueueRunner::CopyReadbackBuffer(int width, int height, Draw::DataFormat srcFormat, Draw::DataFormat destFormat, int pixelStride, uint8_t *pixels) {
 	// TODO: Maybe move data format conversion here, and always read back 8888. Drivers
-	// don't usually provide very optimized implementations.
-	int bpp = Draw::DataFormatSizeInBytes(destFormat);
+	// don't usually provide very optimized conversion implementations, though some do.
+	int bpp = (int)Draw::DataFormatSizeInBytes(destFormat);
 	for (int y = 0; y < height; y++) {
 		memcpy(pixels + y * pixelStride * bpp, readbackBuffer_ + y * width * bpp, pixelStride * bpp);
 	}
